+++ elements/upfront-text/js/settings.js
@@ -96,7 +96,7 @@ define([
 				
 				migrateElementStyle: function(styles, selector) {
 					//replace container class
-					styles = styles.replace(/\.upfront-plain_txt/, ' .plain-text-container');
+					styles = styles.replace(/\.upfront-plain_txt/g, ' .plain-text-container');
 					
 					return styles;
 				},
@@ -110,12 +110,23 @@ define([
 						props[prop.get('name')] = prop.get('value');
 					});
 
+/*
+// Too complex, and can fail in edge cases - check the comment below for the new approach explanation and details
 					if((typeof props.border_color !== "undefined" && props.border_color !== "rgba(0, 0, 0, 0)") && 
 					   (typeof props.border_style !== "undefined" && props.border_style !== "none") && 
 					   (typeof props.border_width !== "undefined" && props.border_width !== 1)) {
 							useBorder = 'yes';
 					}
-					
+*/
+					// Simplify the border usage migration logic: no border _style_ === no border.
+					// Other border-related props might have some defaults crept in or something,
+					// and there's plenty of options there (e.g. for colors). So, keep it real and
+					// check only what _actually_ dictates the border usage in properties.
+					useBorder = typeof props.border_style !== typeof undefined && "none" !== props.border_style
+						? 'yes'
+						: ''
+					;
+
 					if((typeof props.background_color !== "undefined" && props.background_color) || useBorder === 'yes') {
 						usePadding = 'yes';
 					}