+++ scripts/redactor/plugins.js
@@ -1081,6 +1081,16 @@ RedactorPlugins.upfrontColor = function() {
                 background_picker.render();
                 this.$("#tabbackground-content").html(background_picker.el);
 
+                /**
+                 * Position the color pickers so that they are always inside the page
+                 *
+                 */
+                var $dropdown = this.$el.closest(".redactor-dropdown");
+                $dropdown.removeClass( "on-right-border" );
+                if( $(window).width() <= (  this.$("#tabforeground-content").offset().left + this.$("#tabforeground-content").width() ) )
+                    $dropdown.addClass( "on-right-border" );
+
+
                 if (this.current_color) {
                     var color = tinycolor(self.current_color);
                     foreground_picker.$(".upfront-field-color").spectrum("option", "color", self.current_color);